requirements.txt
Django==5.0
django-filter==24.3
djangorestframework-jsonapi==7.1.0
drf-spectacular-jsonapi==0.5.2
gunicorn==23.0.0
mysqlclient==2.2.1
sqlparse==0.5.1
